Abstract date to core
