+++ backend/core/models.py
@@ -1,5 +1,5 @@
 from django.db import models
-
+from django.core.validators import MinValueValidator, MaxValueValidator
 
 class Field(models.Model):
     certainty = models.IntegerField(
@@ -20,3 +20,43 @@ class Field(models.Model):
 
     class Meta:
         abstract = True
+
+class LettercraftDate(models.Model):
+    MIN_YEAR = 400
+    MAX_YEAR = 800
+
+    year_lower = models.IntegerField(
+        validators=[
+            MinValueValidator(MIN_YEAR),
+            MaxValueValidator(MAX_YEAR),
+        ],
+        default=MIN_YEAR,
+        help_text="The earliest possible year for this value",
+    )
+
+    year_upper = models.IntegerField(
+        validators=[
+            MinValueValidator(MIN_YEAR),
+            MaxValueValidator(MAX_YEAR),
+        ],
+        default=MAX_YEAR,
+        help_text="The latest possible year for this value",
+    )
+
+    year_exact = models.IntegerField(
+        null=True,
+        blank=True,
+        validators=[
+            MinValueValidator(MIN_YEAR),
+            MaxValueValidator(MAX_YEAR),
+        ],
+        help_text="The exact year of the value (if known). This will override the values in the lower and upper bounds fields.",
+    )
+
+    class Meta:
+        abstract = True
+
+    def clean(self):
+        if self.year_exact:
+            self.year_lower = self.year_exact
+            self.year_upper = self.year_exact
